more admin routes

+++ back-end/app/controllers/documents_controller.rb
@@ -7,15 +7,25 @@ class DocumentsController < ApplicationController
             render json: {message: 'Failed to create Document'}
         end
      end
+
       def destroy
-        doc = Document.find(params[:id])
-        if doc
+        token = request.headers['Authentication'].split(' ')[1]
+        payload = decode(token) 
+        if payload['user_id'] == 1
+          doc = Document.find(params[:id])
+          if doc
             doc.destroy
             render json: doc
-        else
+          else
             render json: {message: 'Failed to destroy Document'}
+          end
+        else
+        render json: {message: "Unauthorized Action"}
         end
       end
+      
+    private
+
       def doc_params
         params.permit(:name, :order_id,:file)
       end
